% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{gcd}
\alias{gcd}
\title{Calcualte greatest common divisor of two integers}
\usage{
gcd(x, y)
}
\arguments{
\item{x,}{y Two integers}
}
\value{
The greatest common divisor
}
\description{
Calcualte greatest common divisor of two integers
}

