% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{layoutmat}
\alias{layoutmat}
\title{Calcualte the layout matrix}
\usage{
layoutmat(v = 1, equal = TRUE, ratio = equal)
}
\arguments{
\item{v}{a vector each element indicating the total number of plots in one row}

\item{equal}{a logical value indicates whether the layout should be equally distributed}

\item{ratio}{if equal is FALSE, supply division ratio}
}
\value{
A matrix which can be used as graphics::layout argument
}
\description{
Calcualte the layout matrix
}

