% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{pdf_ocr_text}
\alias{pdf_ocr_text}
\alias{pdf_ocr_data}
\title{OCR text extraction}
\usage{
pdf_ocr_text(
  pdf,
  pages = NULL,
  opw = "",
  upw = "",
  language = "eng",
  dpi = 600
)

pdf_ocr_data(
  pdf,
  pages = NULL,
  opw = "",
  upw = "",
  language = "eng",
  dpi = 600
)
}
\arguments{
\item{pdf}{file path or raw vector with pdf data}

\item{pages}{which pages of the pdf file to extract}

\item{opw}{string with owner password to open pdf}

\item{upw}{string with user password to open pdf}

\item{language}{passed to \link[tesseract:tesseract]{tesseract} to specify the
languge of the engine.}

\item{dpi}{resolution to render image that is passed to \link[tesseract:ocr]{tesseract::ocr}.}
}
\description{
Perform OCR text extraction. This requires you have the \code{tesseract} package.
}
\seealso{
Other pdftools: 
\code{\link{pdftools}},
\code{\link{qpdf}},
\code{\link{rendering}}
}
\concept{pdftools}
