% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logo.R
\name{plotlogo}
\alias{plotlogo}
\title{Add Logo Utility}
\usage{
plotlogo(x, y, h, weight_asp = 1)
}
\arguments{
\item{x,y}{position coordinate x and y}

\item{h}{height in unit of lines}

\item{weight_asp}{weight of aspect, default 1, means keeping original aspect}
}
\value{
Invisible NULL
}
\description{
Add Logo Utility
}

