% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\docType{import}
\name{qpdf}
\alias{qpdf}
\alias{pdf_combine}
\alias{reexports}
\alias{pdf_compress}
\alias{pdf_length}
\alias{pdf_split}
\alias{pdf_subset}
\title{qpdf utilities}
\seealso{
Other pdftools: 
\code{\link{pdf_ocr_text}()},
\code{\link{pdftools}},
\code{\link{rendering}}
}
\concept{pdftools}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{qpdf}{\code{\link[qpdf]{pdf_combine}}, \code{\link[qpdf]{pdf_compress}}, \code{\link[qpdf]{pdf_length}}, \code{\link[qpdf]{pdf_split}}, \code{\link[qpdf]{pdf_subset}}}
}}

