% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layout.R
\name{scm}
\alias{scm}
\title{Calcualte smallest common multiple of two integers}
\usage{
scm(x, y)
}
\arguments{
\item{x,}{y Two integers}
}
\value{
The smallest common multiple
}
\description{
Calcualte smallest common multiple of two integers
}

