% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{pdf_render_page}
\alias{pdf_render_page}
\alias{render}
\title{Render PDF to bitmap}
\usage{
pdf_render_page(pdf, pages = 0, dpi = 72, numeric = TRUE, opw = "",
  upw = "")
}
\arguments{
\item{pdf}{file path or raw vector with pdf data}

\item{numeric}{convert raw output to (0-1) real values}

\item{opw}{owner password}

\item{upw}{user password}

\item{page}{which page to render}
}
\description{
Renders a PDF page to a bitmap array which can be written to e.g.
png, jpeg or webp using the respective R packages. This function is only
available if libpoppler was compiled with cairo support.
}
\examples{
\dontrun{
download.file("http://arxiv.org/pdf/1403.2805.pdf", "1403.2805.pdf")
bitmap <- pdf_render_page("1403.2805.pdf")

# save to bitmap formats
png::writePNG(bitmap, "page.png")
jpeg::writeJPEG(bitmap, "page.jpeg")
webp::write_webp(bitmap, "page.webp")

# Higher quality
bitmap <- pdf_render_page("1403.2805.pdf", dpi = 300)
png::writePNG(bitmap, "page.png")

# slightly more efficient
bitmap_raw <- pdf_render_page("1403.2805.pdf", numeric = FALSE)
webp::write_webp(bitmap_raw, "page.webp")
}
}

