% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{pdf_info}
\alias{pdf_attachments}
\alias{pdf_fonts}
\alias{pdf_info}
\alias{pdf_text}
\alias{pdf_toc}
\alias{pdftools}
\alias{poppler_info}
\title{PDF tools}
\usage{
pdf_info(pdf, owner_password = "", user_password = "")

pdf_text(pdf, owner_password = "", user_password = "")

pdf_fonts(pdf, owner_password = "", user_password = "")

pdf_attachments(pdf, owner_password = "", user_password = "")

pdf_toc(pdf, owner_password = "", user_password = "")
}
\arguments{
\item{pdf}{file path or raw vector with pdf data}

\item{owner_password}{string with owner password to open pdf}

\item{user_password}{string with user password to open pdf}
}
\description{
Utilities based on libpoppler for extracting text,
fonts and metadata from a pdf file.
}
\examples{
download.file("http://cran.r-project.org/doc/manuals/r-release/R-exts.pdf",
 "R-exts.pdf", mode = "wb")
pdf_info("R-exts.pdf")
pdf_text("R-exts.pdf")
pdf_fonts("R-exts.pdf")
pdf_files("R-exts.pdf")
}

