% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pdf_info}
\alias{pdf_attachments}
\alias{pdf_fonts}
\alias{pdf_info}
\alias{pdf_text}
\alias{pdf_toc}
\alias{pdftools}
\alias{poppler_info}
\title{PDF tools}
\usage{
pdf_info(pdf, opw = "", upw = "")

pdf_text(pdf, opw = "", upw = "")

pdf_fonts(pdf, opw = "", upw = "")

pdf_attachments(pdf, opw = "", upw = "")

pdf_toc(pdf, opw = "", upw = "")
}
\arguments{
\item{pdf}{file path or raw vector with pdf data}

\item{opw}{string with owner password to open pdf}

\item{upw}{string with user password to open pdf}
}
\description{
Utilities based on libpoppler for extracting text,
fonts and metadata from a pdf file.
}
\examples{
download.file("http://cran.r-project.org/doc/manuals/r-release/R-exts.pdf",
 "R-exts.pdf", mode = "wb")
info <- pdf_info("R-exts.pdf")
text <- pdf_text("R-exts.pdf")
fonts <- pdf_fonts("R-exts.pdf")
files <- pdf_attachments("R-exts.pdf")
}

